changed basic vision boilerplate code to work with Epilogue
