+++ src/main/java/frc/robot/utils/VisionPoseEstimator.java
@@ -6,17 +6,17 @@ package frc.robot.utils;
 
 import edu.wpi.first.epilogue.Logged;
 import edu.wpi.first.epilogue.Logged.Strategy;
-import edu.wpi.first.math.VecBuilder;
-import edu.wpi.first.math.Vector;
-import edu.wpi.first.math.geometry.Pose3d;
 import edu.wpi.first.math.geometry.Transform3d;
-import edu.wpi.first.math.numbers.N3;
 import java.util.ArrayList;
 import java.util.List;
 
 /** Handles pose estimation coming from a single PhotonVision camera. */
 @Logged(strategy = Strategy.OPT_IN)
 public class VisionPoseEstimator {
+  // the latest vision estimate from the pose estimator
+  @Logged(name = "Latest Estimate")
+  private VisionPoseEstimate _latestEstimate = VisionPoseEstimate.noDetectedTags();
+
   /** The camera's NT name. */
   @Logged(name = "Camera Name")
   public final String camName;
@@ -45,8 +45,19 @@ public class VisionPoseEstimator {
   @Logged(name = "Ignore Theta Estimate")
   public final boolean ignoreThetaEstimate = true;
 
-  @Logged(name = "Latest Estimate")
-  private VisionPoseEstimate _latestEstimate = VisionPoseEstimate.noDetectedTags();
+  /** Constants for a single vision pose estimator camera. */
+  public record VisionPoseEstimatorConstants(
+      /** The NT name of the camera. */
+      String camName,
+
+      /** The robot to camera transform */
+      Transform3d robotToCam,
+
+      /** The ambiguity threshold for filtering */
+      double ambiguityThreshold,
+
+      /** The camera's std devs factor. */
+      double cameraStdDevsFactor) {}
 
   /** Builds a new vision pose estimator from a single camera constants. */
   public static VisionPoseEstimator buildFromCamera(VisionPoseEstimatorConstants camConstants) {
@@ -93,56 +104,5 @@ public class VisionPoseEstimator {
     this.robotToCam = robotToCam;
     this.ambiguityThreshold = ambiguityThreshold;
     this.cameraStdDevsFactor = cameraStdDevsFactor;
-  }
-
-  /** Constants for a single vision pose estimator camera. */
-  public record VisionPoseEstimatorConstants(
-      /** The NT name of the camera. */
-      String camName,
-
-      /** The robot to camera transform */
-      Transform3d robotToCam,
-
-      /** The ambiguity threshold for filtering */
-      double ambiguityThreshold,
-
-      /** The camera's std devs factor. */
-      double cameraStdDevsFactor) {}
-  ;
-
-  /** Represents a single vision pose estimate. */
-  @Logged
-  public record VisionPoseEstimate(
-      /** The pose to add into the estimator. */
-      Pose3d pose,
-
-      /** The timestamp of when the frame was taken. (-1 when no tags). */
-      double timestamp,
-
-      /** The ambiguity of this measurement (-1 when no tags or when multi-tag). */
-      double ambiguity,
-
-      /** The detected tag ids in this measurement. */
-      int[] detectedTags,
-
-      /** The average distance from the tag(s) (-1 when no tags). */
-      double avgTagDistance,
-
-      /**
-       * The [xMeters, yMeters, thetaRadians] noise standard deviations of this pose estimate ([-1,
-       * -1, -1] when no tags or invalid).
-       */
-      Vector<N3> stdDevs,
-
-      /** Whether this estimate passed the filter or not. */
-      boolean isValid) {
-    /**
-     * Returns a vision pose estimate that represents an estimate with no detected tags (or camera
-     * was disconnected).
-     */
-    public static final VisionPoseEstimate noDetectedTags() {
-      return new VisionPoseEstimate(
-          new Pose3d(), -1, -1, new int[0], -1, VecBuilder.fill(-1, -1, -1), false);
-    }
   }
 }
